+++ globals.js
@@ -17,7 +17,7 @@ const FILENAME = import.meta.url.split('/').pop();
 // Ενημέρωση για Εκκίνηση Φόρτωσης Αρχείου
 console.log(`[${new Date().toLocaleTimeString()}] 🚀 Φόρτωση: ${FILENAME} ${VERSION} -> Ξεκίνησε`);
 
-// Imports από utils.js
+/* ========================= Imports ========================= */
 import { log, ts, anyTrue, allTrue, rndInt, isDefined, isNonEmptyArray, deepClone, delay, cancel, scheduleSafe } from './utils.js';
 
 /** --- Console Filter (external) Early Install - Start --- */
@@ -238,7 +238,7 @@ if (typeof document !== 'undefined') {
 }
 /** --- UI Utilities - End --- */
 
-// Ενημέρωση για Ολοκλήρωση Φόρτωσης Αρχείου
+/* Ενημέρωση για Ολοκλήρωση Φόρτωσης Αρχείου */
 console.log(`[${new Date().toLocaleTimeString()}] ✅ Φόρτωση: ${FILENAME} ${VERSION} -> Ολοκληρώθηκε`);
 
 // --- End Of File ---